+++ feature/inbox/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/feature/inbox/di/InboxModule.kt
@@ -3,11 +3,17 @@ package com.github.diegoberaldin.raccoonforlemmy.feature.inbox.di
 import com.github.diegoberaldin.raccoonforlemmy.core.architecture.DefaultMviModel
 import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.main.InboxMviModel
 import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.main.InboxViewModel
-import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.messages.InboxMessagesMviModel
-import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.messages.InboxMessagesViewModel
+import com.github.diegoberaldin.raccoonforlemmy.unit.mentions.di.inboxMentionsModule
+import com.github.diegoberaldin.raccoonforlemmy.unit.messages.di.inboxMessagesModule
+import com.github.diegoberaldin.raccoonforlemmy.unit.replies.di.inboxRepliesModule
 import org.koin.dsl.module
 
 val inboxTabModule = module {
+    includes(
+        inboxRepliesModule,
+        inboxMessagesModule,
+        inboxMentionsModule,
+    )
     factory<InboxMviModel> {
         InboxViewModel(
             mvi = DefaultMviModel(InboxMviModel.UiState()),
@@ -16,17 +22,6 @@ val inboxTabModule = module {
             coordinator = get(),
             settingsRepository = get(),
             notificationCenter = get(),
-        )
-    }
-    factory<InboxMessagesMviModel> {
-        InboxMessagesViewModel(
-            mvi = DefaultMviModel(InboxMessagesMviModel.UiState()),
-            identityRepository = get(),
-            siteRepository = get(),
-            messageRepository = get(),
-            coordinator = get(),
-            notificationCenter = get(),
-            settingsRepository = get(),
         )
     }
 }
