+++ feature/inbox/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/feature/inbox/di/Utils.kt
@@ -1,8 +1,5 @@
 package com.github.diegoberaldin.raccoonforlemmy.feature.inbox.di
 
 import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.main.InboxMviModel
-import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.messages.InboxMessagesMviModel
 
 expect fun getInboxViewModel(): InboxMviModel
-
-expect fun getInboxMessagesViewModel(): InboxMessagesMviModel
