+++ settings.gradle.kts
@@ -79,3 +79,4 @@ include(":unit:web")
 include(":unit:zoomableimage")
 include(":unit:replies")
 include(":unit:mentions")
+include(":unit:messages")
