+++ unit/messages/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/messages/di/Utils.kt
@@ -1,0 +1,5 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.messages.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.messages.InboxMessagesMviModel
+
+expect fun getInboxMessagesViewModel(): InboxMessagesMviModel
